Moved class to another package.
