+++ client/src/main/java/org/evosuite/ga/operators/ranking/CrowdingDistance.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with EvoSuite. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.evosuite.ga.metaheuristics.mosa;
+package org.evosuite.ga.operators.ranking;
 
 import java.util.ArrayList;
 import java.util.Collections;
